Add ability to not normalize on camelize
Problem: if you have a string like something_toCamelize you probably
want to keep the last part camelized. With the auto-normalization this
is not possible.

Solution: adding a new option on the method so you can disable
normalization. I also make the normalization more powerful with a little
classi regex.